+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+## Upcoming
+
+#### 5 Jan 2020
+
+- added empty chart to bootstrap new charts
+
 ## v1.0.0-beta.3
+
 #### 15 Dec 2020
 
 - Bugfixing bar chart
 - Added horizontal bar chart
 
 ## v1.0.0-beta.2
+
 #### 01 Dec 2020
 
 Features:
@@ -12,6 +20,7 @@ Features:
 - Matrix plot: using `undefined` for default color in custom mapping
 
 ## v1.0.0-beta.1
+
 #### 01 Dec 2020
 
 Features:
@@ -19,7 +28,9 @@ Features:
 - Added repeated labels for matrix plot
 
 ## v0.0.39
+
 #### 27 Nov 2020
 
 Features:
+
 - Refined Treemap